Support recasting SW:P if haste% increases by at least a certain given amount. Support specifying when to use bloodlust if you use bloodlust_early (changed it from a bool to an int)
For 3.3.0, change Shadow Priests from using Wild Magic Potion to Potion of Speed.

Updated Priest profiles.

git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@3584 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_target.cpp
@@ -198,7 +198,7 @@ void target_t::combat_begin()
       virtual void execute()
       {
         target_t* t = sim -> target;
-        if ( ( sim -> overrides.bloodlust_early && ( sim -> current_time > sim -> overrides.bloodlust_early ) ) ||
+        if ( ( sim -> overrides.bloodlust_early && ( sim -> current_time > ( double ) sim -> overrides.bloodlust_early ) ) ||
              ( t -> health_percentage() < 25 ) ||
              ( t -> time_to_die()       < 60 ) )
         {
